requirements.txt
aiogram==3.0.0b7
python-dotenv==0.19.0
aiosqlite==0.17.0
ldap3==2.9.1